+++ examples/reversed.py
@@ -1,0 +1,13 @@
+s = "abcde"
+reversed_s = reversed(s)
+print(f'str          = "{s}"')
+print(f'reversed_str = "{"".join(reversed_s)}"')
+
+
+print()
+
+
+lst = [1, 2, 3, 4, 5]
+reversed_lst = reversed(lst)
+print("lst          =", lst)
+print("reversed_lst =", list(reversed_lst))