+++ src/lib.rs
@@ -4,9 +4,14 @@ mod enumerate;
 mod input;
 mod iterable;
 pub mod pathlib;
+mod reversed;
+mod sorted;
 pub mod string;
 pub mod time;
 
 pub use crate::enumerate::enumerate;
 pub use crate::input::input;
 pub use crate::iterable::Iterable;
+pub use crate::reversed::reversed;
+pub use crate::reversed::Reversed;
+pub use crate::sorted::sorted;